refactor: extract RefCountGuard to own TU; add counter assertion to BMH test
