+++ ddprof-lib/src/main/cpp/callTraceStorage.cpp
@@ -13,255 +13,9 @@
 #include "vmEntry.h" // For BCI_ERROR constant
 #include "arch.h" // For LP64_ONLY macro and COMMA macro
 #include "guards.h" // For table swap critical sections
-#include "primeProbing.h"
 #include "thread.h"
 #include <string.h>
 #include <atomic>
-#include <time.h>
-
-// RefCountGuard static members
-RefCountSlot RefCountGuard::refcount_slots[RefCountGuard::MAX_THREADS];
-int RefCountGuard::slot_owners[RefCountGuard::MAX_THREADS];
-
-
-// RefCountGuard implementation
-int RefCountGuard::getThreadRefCountSlot() {
-    // Signal-safe collision resolution: use OS::threadId() with semi-random prime step probing
-    ProfiledThread* thrd = ProfiledThread::currentSignalSafe();
-    int tid = thrd != nullptr ? thrd->tid() : OS::threadId();
-
-    // Semi-random prime step probing to eliminate secondary clustering
-    HashProbe probe(static_cast<u64>(tid), MAX_THREADS);
-
-    int slot = probe.slot();
-    for (int i = 0; i < MAX_PROBE_DISTANCE; i++) {
-        // Try to claim this slot atomically
-        int expected = 0;  // Empty slot (no thread ID)
-        if (__atomic_compare_exchange_n(&slot_owners[slot], &expected, tid, false, __ATOMIC_ACQ_REL, __ATOMIC_RELAXED)) {
-            // Successfully claimed the slot
-            return slot;
-        }
-
-        // Check if we already own this slot (reentrant signal delivery).
-        // Return slot + MAX_THREADS so the caller can detect reentrancy and
-        // save/restore active_ptr instead of clearing it on exit.
-        if (__atomic_load_n(&slot_owners[slot], __ATOMIC_ACQUIRE) == tid) {
-            return slot + MAX_THREADS;
-        }
-
-        // Move to next slot using probe
-        if (probe.hasNext()) {
-            slot = probe.next();
-        }
-    }
-
-    // All probing attempts failed - return -1 to indicate failure
-    return -1;
-}
-
-RefCountGuard::RefCountGuard(void* resource) : _active(true), _is_reentrant(false), _my_slot(-1), _saved_ptr(nullptr) {
-    // Get thread refcount slot using signal-safe collision resolution.
-    // Raw slot >= MAX_THREADS means this thread already owns the slot
-    // (reentrant signal delivery); normalize and set _is_reentrant.
-    int raw = getThreadRefCountSlot();
-
-    if (raw == -1) {
-        // Slot allocation failed - refcount guard is inactive
-        _active = false;
-        return;
-    }
-
-    _is_reentrant = (raw >= MAX_THREADS);
-    _my_slot = _is_reentrant ? (raw - MAX_THREADS) : raw;
-
-    if (_is_reentrant) {
-        // Save the outer guard's pointer so we can restore it on exit.
-        _saved_ptr = __atomic_load_n(&refcount_slots[_my_slot].active_ptr, __ATOMIC_ACQUIRE);
-        // Reentrant: increment count BEFORE overwriting active_ptr.
-        // The outer guard already has count>0; by bumping first the scanner
-        // still sees the outer resource (via the unchanged active_ptr) while
-        // count is elevated, preventing a false "outer resource is clear" result.
-        __atomic_fetch_add(&refcount_slots[_my_slot].count, 1, __ATOMIC_RELEASE);
-        __atomic_store_n(&refcount_slots[_my_slot].active_ptr, resource, __ATOMIC_RELEASE);
-    } else {
-        // Non-reentrant (count was 0): store pointer first so the scanner skips
-        // this slot during the activation window (count=0 → treated as inactive).
-        __atomic_store_n(&refcount_slots[_my_slot].active_ptr, resource, __ATOMIC_RELEASE);
-        __atomic_fetch_add(&refcount_slots[_my_slot].count, 1, __ATOMIC_RELEASE);
-    }
-}
-
-RefCountGuard::~RefCountGuard() {
-    if (_active && _my_slot >= 0) {
-        if (_is_reentrant) {
-            // Reentrant: restore outer active_ptr BEFORE decrementing count so the
-            // scanner always observes the outer resource while count is still elevated.
-            __atomic_store_n(&refcount_slots[_my_slot].active_ptr, _saved_ptr, __ATOMIC_RELEASE);
-            __atomic_fetch_sub(&refcount_slots[_my_slot].count, 1, __ATOMIC_RELEASE);
-        } else {
-            // Non-reentrant: decrement first so the scanner skips this slot (count=0),
-            // then clear the pointer and release the slot.
-            __atomic_fetch_sub(&refcount_slots[_my_slot].count, 1, __ATOMIC_RELEASE);
-            __atomic_store_n(&refcount_slots[_my_slot].active_ptr, nullptr, __ATOMIC_RELEASE);
-            __atomic_store_n(&slot_owners[_my_slot], 0, __ATOMIC_RELEASE);
-        }
-    }
-}
-
-RefCountGuard::RefCountGuard(RefCountGuard&& other) noexcept
-    : _active(other._active), _is_reentrant(other._is_reentrant),
-      _my_slot(other._my_slot), _saved_ptr(other._saved_ptr) {
-    other._active = false;
-}
-
-RefCountGuard& RefCountGuard::operator=(RefCountGuard&& other) noexcept {
-    if (this != &other) {
-        // Clean up current state with same ordering as destructor
-        if (_active && _my_slot >= 0) {
-            if (_is_reentrant) {
-                __atomic_store_n(&refcount_slots[_my_slot].active_ptr, _saved_ptr, __ATOMIC_RELEASE);
-                __atomic_fetch_sub(&refcount_slots[_my_slot].count, 1, __ATOMIC_RELEASE);
-            } else {
-                __atomic_fetch_sub(&refcount_slots[_my_slot].count, 1, __ATOMIC_RELEASE);
-                __atomic_store_n(&refcount_slots[_my_slot].active_ptr, nullptr, __ATOMIC_RELEASE);
-                __atomic_store_n(&slot_owners[_my_slot], 0, __ATOMIC_RELEASE);
-            }
-        }
-
-        // Move from other
-        _active       = other._active;
-        _is_reentrant = other._is_reentrant;
-        _my_slot      = other._my_slot;
-        _saved_ptr    = other._saved_ptr;
-
-        // Clear other
-        other._active = false;
-    }
-    return *this;
-}
-
-void RefCountGuard::waitForRefCountToClear(void* table_to_delete) {
-    // Check refcount slots for the table we want to delete
-    //
-    // POINTER-FIRST PROTOCOL GUARANTEES:
-    // - Constructor stores pointer then increments count
-    // - Destructor decrements count then clears pointer
-    // - Scanner checks count first (if 0, slot is inactive)
-    //
-    // TRACE DROP WINDOW (intentional design):
-    // - Scanner can complete on FIRST iteration if all slots have count=0
-    // - Guards in construction (pointer stored, count still 0) are treated as inactive
-    // - Revalidation check in put() detects this race and drops the trace
-    // - This trades a narrow trace-drop window (~10-100ns) for protocol simplicity
-    // - USE-AFTER-FREE IS IMPOSSIBLE: Revalidation prevents table access after deletion
-
-    // PHASE 1: Fast path - spin with pause for short waits (common case)
-    // Expected: refcounts clear within 1-20µs as put() operations complete
-    const int SPIN_ITERATIONS = 100;
-    for (int spin = 0; spin < SPIN_ITERATIONS; ++spin) {
-        bool all_clear = true;
-
-        // Scan all slots (no bitmap optimization, but simpler logic)
-        for (int i = 0; i < MAX_THREADS; ++i) {
-            // CRITICAL: Check count FIRST (pointer-first protocol)
-            uint32_t count = __atomic_load_n(&refcount_slots[i].count, __ATOMIC_ACQUIRE);
-            if (count == 0) {
-                continue;  // Slot inactive, skip it
-            }
-
-            // Count > 0, so slot is active - check which table it protects
-            void* table = __atomic_load_n(&refcount_slots[i].active_ptr, __ATOMIC_ACQUIRE);
-            if (table == table_to_delete) {
-                all_clear = false;
-                break;
-            }
-        }
-
-        if (all_clear) {
-            return; // Fast path success - refcounts cleared quickly
-        }
-        spinPause(); // CPU pause instruction, ~10-50 cycles
-    }
-
-    // PHASE 2: Slow path - async-signal-safe sleep for blocked thread case
-    const int MAX_WAIT_ITERATIONS = 5000;
-    struct timespec sleep_time = {0, 100000}; // 100 microseconds
-
-    for (int wait_count = 0; wait_count < MAX_WAIT_ITERATIONS; ++wait_count) {
-        bool all_clear = true;
-
-        for (int i = 0; i < MAX_THREADS; ++i) {
-            uint32_t count = __atomic_load_n(&refcount_slots[i].count, __ATOMIC_ACQUIRE);
-            if (count == 0) {
-                continue;
-            }
-
-            void* table = __atomic_load_n(&refcount_slots[i].active_ptr, __ATOMIC_ACQUIRE);
-            if (table == table_to_delete) {
-                all_clear = false;
-                break;
-            }
-        }
-
-        if (all_clear) {
-            return; // Slow path success
-        }
-
-        // nanosleep is POSIX async-signal-safe and does not call malloc
-        nanosleep(&sleep_time, nullptr);
-    }
-
-    // If we reach here, some refcounts didn't clear in time — proceed anyway
-    // to avoid an infinite hang; the caller is the only thread that frees the
-    // buffer, so proceeding is safe in practice.
-    Log::warn("waitForRefCountToClear: timeout waiting for %p, proceeding", table_to_delete);
-}
-
-void RefCountGuard::waitForAllRefCountsToClear() {
-    // PHASE 1: Fast path - spin with pause for short waits
-    const int SPIN_ITERATIONS = 100;
-    for (int spin = 0; spin < SPIN_ITERATIONS; ++spin) {
-        bool any_refcounts = false;
-
-        for (int i = 0; i < MAX_THREADS; ++i) {
-            uint32_t count = __atomic_load_n(&refcount_slots[i].count, __ATOMIC_ACQUIRE);
-            if (count > 0) {
-                any_refcounts = true;
-                break;
-            }
-        }
-
-        if (!any_refcounts) {
-            return; // Fast path success
-        }
-        spinPause();
-    }
-
-    // PHASE 2: Slow path - async-signal-safe sleep
-    const int MAX_WAIT_ITERATIONS = 5000;
-    struct timespec sleep_time = {0, 100000}; // 100 microseconds
-
-    for (int wait_count = 0; wait_count < MAX_WAIT_ITERATIONS; ++wait_count) {
-        bool any_refcounts = false;
-
-        for (int i = 0; i < MAX_THREADS; ++i) {
-            uint32_t count = __atomic_load_n(&refcount_slots[i].count, __ATOMIC_ACQUIRE);
-            if (count > 0) {
-                any_refcounts = true;
-                break;
-            }
-        }
-
-        if (!any_refcounts) {
-            return; // Slow path success
-        }
-
-        nanosleep(&sleep_time, nullptr);
-    }
-
-    // If we reach here, some refcounts didn't clear in time
-}
-
 
 static const u64 OVERFLOW_TRACE_ID = 0x7fffffffffffffffULL;  // Max 64-bit signed value
