+++ ddprof-test/src/test/java/com/datadoghq/profiler/metadata/BoundMethodHandleMetadataSizeTest.java
@@ -8,6 +8,8 @@ import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
 
+import java.util.Map;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assumptions.assumeFalse;
 
@@ -31,6 +33,8 @@ public class BoundMethodHandleMetadataSizeTest extends AbstractProfilerTest {
         assertTrue(x != 0);
         stopProfiler();
         verifyEvents("datadog.MethodSample");
+        Map<String, Long> counters = profiler.getDebugCounters();
+        assertFalse(counters.isEmpty(), "profiler debug counters must not be empty after BoundMethodHandle workload");
     }
